Merge from Jetbrains

+++ java/java-tests/testData/inspection/inefficientStreamCount/afterStreamCount.java
@@ -5,6 +5,6 @@ import java.util.Arrays;
 class Test {
   long cnt() {
       /*count*/
-      return (long) Arrays.asList('d', 'e', 'f')./*stream*/size()/*after*/;
+      return Arrays.asList('d', 'e', 'f')./*stream*/size()/*after*/;
   }
 }
